new bot/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
Django==4.0.6
django-jazzmin==2.5.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tornado==6.2
tzdata==2022.1
tzlocal==4.2
